Finallygit add .
